array list of points created

+++ app/src/main/java/com/kristijan_pajtasev/assignment03/LocationPoint.java
@@ -1,10 +1,10 @@
 package com.kristijan_pajtasev.assignment03;
 
 public class LocationPoint {
-    private float altitude, latitude, longitude;
+    private double altitude, latitude, longitude;
     private long time;
 
-    public LocationPoint(float altitude, float latitude, float longitude, long time) {
+    public LocationPoint(double altitude, double latitude, double longitude, long time) {
         this.altitude = altitude;
         this.latitude = latitude;
         this.longitude = longitude;
